fix: add author ownership check for updating and deleting blogs

+++ server/controllers/blogs.controller.js
@@ -35,12 +35,19 @@ export const createBlog = asyncHandler(async (req, res, next) => {
 
 // Update a blog
 export const updateBlog = asyncHandler(async (req, res, next) => {
-    const blog = await Blog.findByIdAndUpdate(
+    let blog = await Blog.findById(req.params.id);
+    if (!blog) return next(new errorHandler("Blog not found", 404));
+
+    if (blog.author.toString() !== req.user._id.toString()) {
+        return next(new errorHandler("Unauthorized", 403));
+    }
+
+    blog = await Blog.findByIdAndUpdate(
         req.params.id,
         req.body,
         { new: true, runValidators: true }
     );
-    if (!blog) return next(new errorHandler("Blog not found", 404));
+
     res.status(200).json({
         success: true,
         responseData: blog,
@@ -49,8 +56,15 @@ export const updateBlog = asyncHandler(async (req, res, next) => {
 
 // Delete a blog
 export const deleteBlog = asyncHandler(async (req, res, next) => {
-    const blog = await Blog.findByIdAndDelete(req.params.id);
+    let blog = await Blog.findById(req.params.id);
     if (!blog) return next(new errorHandler("Blog not found", 404));
+
+    if (blog.author.toString() !== req.user._id.toString()) {
+        return next(new errorHandler("Unauthorized", 403));
+    }
+
+    blog = await Blog.findByIdAndDelete(req.params.id);
+
     res.status(200).json({
         success: true,
         responseData: blog,
